requirements.txt
ccxt==4.2.1
pandas==2.2.0
python-telegram-bot==20.8
python-dotenv==1.0.1
requests==2.31.0
